+++ exporter.cpp
@@ -1,0 +1,5 @@
+#include<iostream>
+
+void getExport(){
+	std::cout<<"Export feature coming soon!";
+}
